requirements.txt
django==5.0.2 
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-environ==0.11.2
djangorestframework-simplejwt==4.4.0
django-model-utils==4.0.0
